requirements.txt
django==2.2
djangorestframework==3.11.0
PyJWT==1.7.1
djangorestframework-jwt==1.11.0
django-filters==0.2.1
django-filter==2.3.0